tests: Improve fixtures typing #TidyFridy
