+++ tests/tests/scenario.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class Scenario:
+    project: int
+    op_study: int
+    scenario: int
+    infra: int
+    timetable: int
