+++ tests/tests/test_rtk_sync.py
@@ -1,7 +1,7 @@
 import subprocess
 from pathlib import Path
 
-FRONT_FOLDER = Path(__file__).parent.parent.parent / "front"
+FRONT_FOLDER = Path(__file__).parents[2] / "front"
 MIDDLEWARE_FOLDER = FRONT_FOLDER / "src" / "common" / "api"
 MIDDLEWARE_FILES = {
     "api": MIDDLEWARE_FOLDER / "osrdEditoastApi.ts",
