+++ tests/tests/test_timetable.py
@@ -1,9 +1,10 @@
 import requests
 
+from .scenario import Scenario
 from .services import API_URL
 
 
-def test_get_timetable(dummy_scenario):
+def test_get_timetable(dummy_scenario: Scenario):
     r = requests.get(
         API_URL
         + f"projects/{dummy_scenario.project}/studies/{dummy_scenario.op_study}/scenarios/{dummy_scenario.scenario}/"
